Tester push
I have pushed some changes in master of Muzzler.Windows also. feel free
to revert that one.

I am pushing this with no fix but it clearly shows the problem. Well not
clearly, but for some reason our camera scaling is off, so either the
camera or maybe /projection matrix/. Changes:

-Hard coded many test cases
-Clip still fails to give me a proper clip (Dims wasn't initialized when
I'm getting it)
-Deleted USELESS lines in Quad, as well as added a line that may be
removed later
-Shadows and resizing are messed up, but that wasn't a priority

To get a proper look on how the scaling should be, refer to
2DRendering@D3D project in samples/window in Dropbox.